requirements.txt
altgraph==0.17.4
ffmpeg-python==0.2.0
future==1.0.0
install==1.3.5
packaging==24.1
pyasn1==0.6.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.1
PyQt6-sip==13.6.0
rsa==4.9
setuptools==70.1.1
